Move DebugConsole to editor and rename
